feat: adds VANOS map for 41.2/41.3 @ 0x2DC
adds "VANOS - Minimum RPM (Closed Throttle)" map to 41.2 and 41.3. The stock value at address 0x2DC is 0x7D or 125 decimal. The expression is (x * 32) resulting in a stock value of 4000. The units are RPMs. If the current RPM is greater than this value, VANOS will remain activated with closed throttle as long as "VANOS - Min. Load (Light Throttle) (MT)/(AT)" load conditions are met.